Working on sqlite integration

+++ src/models.rs
@@ -3,14 +3,14 @@ use crate::schema::subscriptions;
 
 #[derive(Queryable)]
 pub struct Subscription {
-    pub id: u32,
-    pub user_id: u32,
-    pub created_at: u32,
+    pub id: i32,
+    pub user_id: i64,
+    pub created_at: i32,
 }
 
 #[derive(Insertable)]
 #[diesel(table_name = subscriptions)]
 pub struct NewSubscription {
-    pub user_id: i32,
+    pub user_id: i64,
     pub created_at: i32,
 }
